+++ klippy/extras/display/menu.py
@@ -5,9 +5,7 @@
 # Copyright (C) 2018  Janar Sööt <janar.soot@gmail.com>
 #
 # This file may be distributed under the terms of the GNU GPLv3 license.
-import os, ConfigParser, logging
-import sys, ast, re
-import klippy
+import os, logging, sys, ast, re
 
 
 class error(Exception):
@@ -954,10 +952,9 @@ class MenuManager:
                                         desc=self.cmd_DO_help)
 
         # Parse local config file in same directory as current module
-        fileconfig = ConfigParser.RawConfigParser()
+        pconfig = self.printer.lookup_object('configfile')
         localname = os.path.join(os.path.dirname(__file__), 'menu.cfg')
-        fileconfig.read(localname)
-        localconfig = klippy.ConfigWrapper(self.printer, fileconfig, {}, None)
+        localconfig = pconfig.read_config(localname)
 
         # Load items from local config
         self.load_menuitems(localconfig)
